refactor(macro): move agent assets from server/ to agent/
Moves tools, sdk, transforms, and workspace-template out of server/ into
a new agent/ layer for the macro plugin. Updates server/index.ts to use
../agent/tools/ imports and ../agent/ provisioning URLs. Also extracts
dataCatalogPlugin agent code into its own agent/ layer.
