+++ packages/workspace/templates/plugin/server/index.ts
@@ -1,12 +1,12 @@
-import type { AgentTool } from "@boring/workspace"
 import { defineServerPlugin } from "@boring/workspace/app/server"
 import { SAMPLE_PLUGIN_ID } from "../shared/constants"
+import { createSampleTools } from "../agent/tools/sampleTools"
 
 export function createSampleServerPlugin() {
   return defineServerPlugin({
     id: SAMPLE_PLUGIN_ID,
     label: "Sample",
-    agentTools: [] satisfies AgentTool[],
+    agentTools: createSampleTools(),
     systemPrompt: "## Sample Plugin",
   })
 }
